pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "arviz>=0.21.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pymc>=5.22.0",
    "pymc-marketing>=0.13.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
]
